Add SVG elements and helper for rotation.
The useRotation hook allows for rotating an arbitrary child element of
an SVGElement. This allows for rotating just pieces of an element,
rather than having to rotate the entire element.

You'll need to make sure that transform-box and transform-origin CSS
properties are set.
For these two, #rotateHandle is used -- It can be set in inkscape on a
group.

Then use the following CSS:
  transform-box: fill-box;
  transform-origin: 50% 50%;
}